+++ server/src/main/java/org/opensearch/extensions/ExtensionsOrchestrator.java
@@ -34,12 +34,15 @@ import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.common.io.FileSystemUtils;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.settings.Settings;
+import org.opensearch.common.settings.SettingsModule;
 import org.opensearch.common.transport.TransportAddress;
 import org.opensearch.discovery.InitializeExtensionsRequest;
 import org.opensearch.discovery.InitializeExtensionsResponse;
 import org.opensearch.extensions.ExtensionsSettings.Extension;
 import org.opensearch.extensions.rest.RegisterRestActionsRequest;
 import org.opensearch.extensions.rest.RestActionsRequestHandler;
+import org.opensearch.extensions.settings.CustomSettingsRequestHandler;
+import org.opensearch.extensions.settings.RegisterCustomSettingsRequest;
 import org.opensearch.index.IndexModule;
 import org.opensearch.index.IndexService;
 import org.opensearch.index.IndicesModuleRequest;
@@ -70,6 +73,7 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
     public static final String REQUEST_EXTENSION_CLUSTER_STATE = "internal:discovery/clusterstate";
     public static final String REQUEST_EXTENSION_LOCAL_NODE = "internal:discovery/localnode";
     public static final String REQUEST_EXTENSION_CLUSTER_SETTINGS = "internal:discovery/clustersettings";
+    public static final String REQUEST_EXTENSION_REGISTER_CUSTOM_SETTINGS = "internal:discovery/registercustomsettings";
     public static final String REQUEST_EXTENSION_REGISTER_REST_ACTIONS = "internal:discovery/registerrestactions";
     public static final String REQUEST_EXTENSION_REGISTER_TRANSPORT_ACTIONS = "internal:discovery/registertransportactions";
     public static final String REQUEST_OPENSEARCH_NAMED_WRITEABLE_REGISTRY = "internal:discovery/namedwriteableregistry";
@@ -90,6 +94,7 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
         REQUEST_EXTENSION_CLUSTER_SETTINGS,
         REQUEST_EXTENSION_ACTION_LISTENER_ON_FAILURE,
         REQUEST_EXTENSION_REGISTER_REST_ACTIONS,
+        REQUEST_EXTENSION_REGISTER_SETTINGS,
         CREATE_COMPONENT,
         ON_INDEX_MODULE,
         GET_SETTINGS
@@ -110,6 +115,7 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
     // A map of extension uniqueId to full extension details used for node transport here and in the RestActionsRequestHandler
     Map<String, DiscoveryExtension> extensionIdMap;
     RestActionsRequestHandler restActionsRequestHandler;
+    CustomSettingsRequestHandler customSettingsRequestHandler;
     TransportService transportService;
     ClusterService clusterService;
     ExtensionNamedWriteableRegistry namedWriteableRegistry;
@@ -145,16 +151,19 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
      * Lists/maps of extensions have already been initialized but not yet populated.
      *
      * @param restController  The RestController on which to register Rest Actions.
+     * @param settingsModule
      * @param transportService  The Node's transport service.
      * @param clusterService  The Node's cluster service.
      */
     public void initializeServicesAndRestHandler(
         RestController restController,
+        SettingsModule settingsModule,
         TransportService transportService,
         ClusterService clusterService
     ) {
         this.restActionsRequestHandler = new RestActionsRequestHandler(restController, extensionIdMap, transportService);
         this.listenerHandler = new ExtensionActionListenerHandler(listener);
+        this.customSettingsRequestHandler = new CustomSettingsRequestHandler(settingsModule);
         this.transportService = transportService;
         this.clusterService = clusterService;
         registerRequestHandler();
@@ -168,6 +177,14 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
             false,
             RegisterRestActionsRequest::new,
             ((request, channel, task) -> channel.sendResponse(restActionsRequestHandler.handleRegisterRestActionsRequest(request)))
+        );
+        transportService.registerRequestHandler(
+            REQUEST_EXTENSION_REGISTER_CUSTOM_SETTINGS,
+            ThreadPool.Names.GENERIC,
+            false,
+            false,
+            RegisterCustomSettingsRequest::new,
+            ((request, channel, task) -> channel.sendResponse(customSettingsRequestHandler.handleRegisterCustomSettingsRequest(request)))
         );
         transportService.registerRequestHandler(
             REQUEST_EXTENSION_CLUSTER_STATE,
@@ -339,16 +356,13 @@ public class ExtensionsOrchestrator implements ReportingService<PluginsAndModule
      * @throws Exception if the request is not handled properly.
      */
     TransportResponse handleExtensionRequest(ExtensionRequest extensionRequest) throws Exception {
-        // Read enum
         switch (extensionRequest.getRequestType()) {
             case REQUEST_EXTENSION_CLUSTER_STATE:
                 return new ClusterStateResponse(clusterService.getClusterName(), clusterService.state(), false);
             case REQUEST_EXTENSION_LOCAL_NODE:
-                LocalNodeResponse localNodeResponse = new LocalNodeResponse(clusterService);
-                return localNodeResponse;
+                return new LocalNodeResponse(clusterService);
             case REQUEST_EXTENSION_CLUSTER_SETTINGS:
-                ClusterSettingsResponse clusterSettingsResponse = new ClusterSettingsResponse(clusterService);
-                return clusterSettingsResponse;
+                return new ClusterSettingsResponse(clusterService);
             default:
                 throw new IllegalArgumentException("Handler not present for the provided request");
         }